JS help needed. Little css also needed
Need to make the collapsible navbar hidden on the blur event and also need an optimized code for the showspecs() function.

As for the css, any styling done on the smallest media query breakpoint present in the code does not work, making me unable to style my carousel. Please HELP